requirements.txt
echo "python-docx==0.8.11
pdf2docx==0.5.6
Pillow==9.3.0
PyPDF2==3.0.0
fpdf==1.7.2" > requirements.txt
